api/src/requirements.txt
click==7.1.2
Flask==1.1.2
flask-marshmallow==0.12.0
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.23.0
pyodbc==4.0.30
python-decouple==3.3
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
